Include LDAP reset in running the tests.
Since we know that the tests won't work if the LDAP is setup wrongly,
force a reset before each run. This doesn't address the issue of things
in the LDAP ending up out of sync as a result of previous tests, but
is a useful step in the right direction for now.

Change-Id: I7be5e0a4e3cd6001a913b1f7c37879e724c8b256