Let some blueshirts see (limited) details of all competitors
This adds support for a 'mentors-extra' group of blueshirts who
have access to see most of the details for any other user. This
access includes their full name, but does not include their email
address.

The primary use case for this is the reception desk at the competition,
where we'll need to check the details against the people that turn
up, and (in particular) their media consent status.

Change-Id: I8175a526601ef2e5638f12aa9d47016a1acb4be0